Since last upload, finished the playerArrows, adding collision between them and enemyies and walls, added a way to remove them from the game as they're destroyed using an iterator to go through the vector of arrows and remove any that have been set to be destroyed. improved the arrow movement . Also created second enemy type, the medusa and did all of it's animation,collision,etc. currently has an attack that correctly tells the player that it has been turned to stone, but need to code the player actually being stone (currently sets the bool, but the player isn't checking that bool). Created fake walls that will be affected by items and use kinametic bodies instead of static so forces can be applied to them. they use a darker version of the wall texture to make it clear they are a bit different.
Improved the player jump by changing when it can be triggered and adding sensors to bottom, left and right of the player so they can be used as a more effective method of detcting when player can jump (allows me to filter out head collisions with walls.) sensors on left and right with collision allows jump, cause it acts as a type of wall climbing, which i wanted to implement anyway.
Changed the majority of pointers used (e.g for the vectors holding enemies, walls and projectiles) to use std::shared_ptr instead, becasue after research, I found that shared_ptr's handle the memory and freeing up the pointer once it's not being used anymore automatically, whereas normals pointers this has to be done manually, which would have been alot of extra methods and code.